# docs/openapi/season-service-v1.yaml
openapi: 3.1.0
info:
  title: ACM Season Service API
  description: API documentation for ACM Season Service
  license:
    name: Apache 2.0
    url: https://springdoc.org
  version: 1.0.0
servers:
- url: http://localhost
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/v1/field-logs/{id}:
    get:
      tags:
      - field-log-controller
      summary: Get field log detail
      description: Get field log detail if it belongs to a season of current farmer
      operationId: getFieldLog
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFieldLogResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFieldLogResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFieldLogResponse"
        "404":
          description: Field log not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFieldLogResponse"
    put:
      tags:
      - field-log-controller
      summary: Update field log
      description: Update field log content while season is still open
      operationId: updateFieldLog
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFieldLogRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFieldLogResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFieldLogResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFieldLogResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFieldLogResponse"
        "404":
          description: Field log not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFieldLogResponse"
    delete:
      tags:
      - field-log-controller
      summary: Delete field log
      description: Delete a field log if allowed by season constraints
      operationId: deleteFieldLog
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "404":
          description: Field log not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
...
  /api/v1/field-logs:
    get:
      tags:
      - field-log-controller
      summary: List field logs
      description: List field logs for a given season of the current farmer
      operationId: listFieldLogs
      parameters:
      - name: seasonId
        in: query
        description: Season ID (required)
        required: true
        schema:
          type: integer
          format: int32
      - name: type
        in: query
        description: Filter by log type
        required: false
        schema:
          type: string
      - name: q
        in: query
        description: Search text in notes (min 2 chars)
        required: false
        schema:
          type: string
      - name: from
        in: query
        description: From date (yyyy-MM-dd)
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        description: To date (yyyy-MM-dd)
        required: false
        schema:
          type: string
          format: date
      - name: page
        in: query
        description: Page index (0-based)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: sort
        in: query
        description: Sort field
        required: false
        schema:
          type: string
          default: "logDate,desc"
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseFieldLogResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseFieldLogResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseFieldLogResponse"
        "404":
          description: Season not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseFieldLogResponse"
    post:
      tags:
      - field-log-controller
      summary: Create field log
      description: Create a new field log linked to a season
      operationId: createFieldLog
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFieldLogRequest"
        required: true
      responses:
        "200":
          description: Success
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFieldLogResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFieldLogResponse"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFieldLogResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFieldLogResponse"
        "404":
          description: Season not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFieldLogResponse"
...
components:
  schemas:
...
    UpdateFieldLogRequest:
      type: object
      properties:
        logDate:
          type: string
          format: date
        logType:
          type: string
          maxLength: 100
          minLength: 0
        notes:
          type: string
          maxLength: 4000
          minLength: 0
      required:
      - logDate
      - logType
    ApiResponseFieldLogResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          $ref: "#/components/schemas/FieldLogResponse"
        data:
          $ref: "#/components/schemas/FieldLogResponse"
        generatedAt:
          type: string
    FieldLogResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        seasonId:
          type: integer
          format: int32
        seasonName:
          type: string
        logDate:
          type: string
          format: date
        logType:
          type: string
        notes:
          type: string
        createdByUserId:
          type: integer
          format: int64
        createdByUsername:
          type: string
        createdByDisplayName:
          type: string
        createdByType:
          type: string
        canEdit:
          type: boolean
        canDelete:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    CreateFieldLogRequest:
      type: object
      properties:
        seasonId:
          type: integer
          format: int32
        logDate:
          type: string
          format: date
        logType:
          type: string
          maxLength: 100
          minLength: 0
        notes:
          type: string
          maxLength: 4000
          minLength: 0
      required:
      - logDate
      - logType
...
    ApiResponsePageResponseFieldLogResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          $ref: "#/components/schemas/PageResponseFieldLogResponse"
        data:
          $ref: "#/components/schemas/PageResponseFieldLogResponse"
        generatedAt:
          type: string
    PageResponseFieldLogResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FieldLogResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    ApiResponseVoid:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          type: object
        data:
          type: object
        generatedAt:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT